Refactor into hz package
