+++ README.md
@@ -1,1 +1,7 @@
 # HZ
+
+This package provides the `hz` module. Run the application using:
+
+```bash
+python -m hz
+```
